Re-organize the structure of the project
