+++ lib/pages/search_page/get_community.dart
